reorganization of sugarscape model
