+++ mesa/examples/advanced/sugarscape_g1mt/app.py
@@ -1,7 +1,13 @@
+import sys
+import os.path
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../../../../')))
+
+
 import numpy as np
 import solara
 from matplotlib.figure import Figure
 from mesa.visualization import SolaraViz, make_plot_measure
+
 from sugarscape_g1mt.model import SugarscapeG1mt
 from sugarscape_g1mt.trader_agents import Trader
 
@@ -14,19 +20,19 @@ def SpaceDrawer(model):
             "trader": {"x": [], "y": [], "c": "tab:red", "marker": "o", "s": 10},
         }
 
-        for content, (i, j) in g.coord_iter():
-            for agent in content:
-                if isinstance(agent, Trader):
-                    layers["trader"]["x"].append(i)
-                    layers["trader"]["y"].append(j)
-                else:
-                    # Don't visualize resource with value <= 1.
-                    layers["sugar"][i][j] = (
-                        agent.sugar_amount if agent.sugar_amount > 1 else np.nan
-                    )
-                    layers["spice"][i][j] = (
-                        agent.spice_amount if agent.spice_amount > 1 else np.nan
-                    )
+        for agent in g.all_cells.agents:
+            i, j = agent.cell.coordinate
+            if isinstance(agent, Trader):
+                layers["trader"]["x"].append(i)
+                layers["trader"]["y"].append(j)
+            else:
+                # Don't visualize resource with value <= 1.
+                layers["sugar"][i][j] = (
+                    agent.sugar_amount if agent.sugar_amount > 1 else np.nan
+                )
+                layers["spice"][i][j] = (
+                    agent.spice_amount if agent.spice_amount > 1 else np.nan
+                )
         return layers
 
     fig = Figure()
