--- mesa/examples/advanced/sugarscape_g1mt/requirements.txt
@@ -1,6 +1,0 @@
-jupyter
-mesa~=2.0
-numpy
-matplotlib
-networkx
-pandas
