+++ mesa/examples/advanced/sugarscape_g1mt/sugar-map.txt
