+++ mesa/examples/advanced/sugarscape_g1mt/tests.py
@@ -1,11 +1,7 @@
-import random
-
 import numpy as np
 from scipy import stats
-from sugarscape_g1mt.model import SugarscapeG1mt, flatten
-from sugarscape_g1mt.trader_agents import Trader
-
-random.seed(1)
+from .model import SugarscapeG1mt, flatten
+from .agents import Trader
 
 
 def check_slope(y, increasing):
@@ -19,7 +15,7 @@ def check_slope(y, increasing):
 def test_decreasing_price_variance():
     # The variance of the average trade price should decrease over time (figure IV-3)
     # See Growing Artificial Societies p. 109.
-    model = SugarscapeG1mt()
+    model = SugarscapeG1mt(42)
     model.datacollector._new_model_reporter(
         "price_variance",
         lambda m: np.var(
